feat(sat): cadical-faithful rephase & target phases; ProbSAT walk
Replace the 8-line inline polarity inverter in restart() (off by
default) and the three dead parallel implementations (rephasing.rs,
target_phase.rs, local_search.rs) with a port of cadical's actual
mechanism (rephase.cpp / backtrack.cpp / walk.cpp):

- propagate maintains no_conflict_until: full trail on a clean
  fixpoint, pre-decision-level prefix on a conflict
  (control[level].trail); stable_conflicts counted where cadical
  counts stabconflicts.
- Every phase-saving backtrack routes through update_target_and_best
  (not just restart): refreshes phases.target (stable-mode decision
  polarity, opts.target=1) and phases.best from the largest
  conflict-free prefix. copy_phases refreshes trail-resident saved
  phases first, so the snapshot is bit-for-bit what cadical (which
  saves at assign time) would have copied.
- rephase fires from BOTH search loops on cadical's arithmetic schedule
  (rephaseint x (total+1), conflict- or stable-conflict-keyed per
  rephase mode), after restart in the same iteration precedence, and
  replays the mode-dependent strategy cycles incl. best/original/
  inverted/flipping/random and a real ProbSAT walk. A best rephase
  re-arms best_assigned at the first post-rephase conflict.
- solver/walk.rs: ProbSAT round over irredundant clauses - occurrence
  lists, Balint CB-value scoring table (size-interpolated every second
  round), roulette literal pick by break count, tick budget at
  walkeffort=80 permille of search ticks since last walk (clamped to
  cadical's walkmaxeff), level-0 variables never flipped, best
  assignment written back into saved phases; interrupt polled every
  1024 flips.
- decision_polarity = cadical decide_phase(idx, target): forced
  (theory-deterministic) > target (stable or target=2) > saved phase.
- Config: rephase (0/1/2, default 1), rephase_interval (u64, 1000),
  target (1), walk/walk_nonstable/walk_effort (cadical defaults);
  old rephase_interval: u32 = 0 (off) removed.
- no_conflict_until clamped on every trail shrink incl. probe
  backtrack, pop, and incremental forget; rephase state reset in
  reset(); init_rephase_limits on every solve (cadical
  init_search_limits).

Measured (satcomp2025 main_easy_mid informative band, instances
solving 0.5-50s): total time -49%, 4 baseline timeouts now solved
(3 sat + 1 unsat, consistent). Z3 differential parity unchanged:
167 decisive, 0 disagreements. Full workspace suite 9925 passed;
clippy/fmt/doc clean.

+++ oxiz-sat/examples/cnf_bench.rs
@@ -77,7 +77,7 @@ fn main() {
         }
     };
     if let Ok(v) = std::env::var("REPHASE")
-        && let Ok(n) = v.parse::<u32>()
+        && let Ok(n) = v.parse::<u64>()
     {
         base_config.rephase_interval = n;
     }
